Add jumphost (#3)
The commit adds a jumphost module. It creates an autoscaling group with
an instance that is supposed to serve as the jumphost.

The ASG has two lifecycle hooks - at launch and terminate event.
The hooks are supposed to update DNS so the jumphost is available via a
fixed name. The lifecycle hooks run a lambda function.
At this moment the lambda does nothing but completes the lifecycle hook
action.