Prevent undesired lambda code changes
If the lambda code changes, the archive name will change, too. This
ensures all lambda related stuff will be changed accordingly. No need to
trigger a chnage when local checksump changed. It can change because of
a number of reasons. Some depenendcy patch-level version can be changed.
We don't want to reupload the working lambda then.