Pass extra policies as a map
This is a feature/limitation of Terraform itself.
If used with for_each, a list must be know at the plan stage. It's not
because ARNs are calculated. As a workaround the module passes a map
with known keys where a value is a calculated ARN. The policy name aka
te map key isn't used in the instance_profile module.