Migrate Terraform registry to infrahouse/registry/aws module
The InfraHouse Terraform registry a.k.a registry.infrahouse.com was
created as an ECS service "terraform-registry".

Some time ago, I create the "infrahouse/ecs/aws" module and now am
migrating to it.

I deployed the registry with the "infrahouse/ecs/aws" module. It is
available now under registry.infrahouse.com name.

The new registry uses service name terraform-registry-2. I had to change
it because the service name is used for the capacity provider name (it
was a mistake btw) and there was a name conflict.

The old registry resource are still available under the registry-old.infrahouse.com name.

The new repo uses a new S3 bucket for modules, however resues DynamoDB
tables which are terraformed now.